requirements.txt
opencv-python==4.10.0.84
numpy==2.1.0
pillow==10.4.0
imutils==0.5.4